refactor to use login manager
